+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Missing empty BOT when writing EOT.
+- Fixed accessing settings for PillowEncoder.
 
 ## [0.14.0] - 2023-11-29