+++ CHANGELOG.md
@@ -7,8 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Fallback to EOT when overflowing BOT.
+
 ### Fixed
 
+- Missing empty BOT when writing EOT.
 - Fixed accessing settings for PillowEncoder.
 
 ## [0.14.0] - 2023-11-29