+++ CHANGELOG.md
@@ -7,9 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.15.0] - 2023-11-30
+
 ### Added
 
 - Fallback to EOT when overflowing BOT.
+- Use AvailableTransferSyntaxUID if provided to determine transfer syntax to use when opening DICOM Web instances.
 
 ### Fixed
 
@@ -226,7 +229,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial release of wsidicom
 
-[Unreleased]: https://github.com/imi-bigpicture/wsidicom/compare/0.12.0..HEAD
+[Unreleased]: https://github.com/imi-bigpicture/wsidicom/compare/0.15.0..HEAD
+[0.15.0]: https://github.com/imi-bigpicture/wsidicom/compare/v0.14.0..v0.15.0
+[0.14.0]: https://github.com/imi-bigpicture/wsidicom/compare/v0.13.0..v0.14.0
+[0.13.0]: https://github.com/imi-bigpicture/wsidicom/compare/v0.12.0..v0.13.0
 [0.12.0]: https://github.com/imi-bigpicture/wsidicom/compare/v0.11.0..v0.12.0
 [0.11.0]: https://github.com/imi-bigpicture/wsidicom/compare/v0.10.0..v0.11.0
 [0.10.0]: https://github.com/imi-bigpicture/wsidicom/compare/v0.9.0..v0.10.0